command tests and handling exception

+++ Domain/Commands/DeleteDogCommand.cs
@@ -18,6 +18,7 @@ namespace Domain.Commands
     public class DeleteDogCommandResult
     {
         public bool DeletingIsSuccessful { get; init; }
+        public string Comment { get; init; }
     }
 
     internal class DeleteDogCommandHandler : IRequestHandler<DeleteDogCommand, DeleteDogCommandResult>
@@ -31,6 +32,15 @@ namespace Domain.Commands
         public async Task<DeleteDogCommandResult> Handle(DeleteDogCommand request, CancellationToken cancellationToken = default)
         {
             Dog dog = await _dbContext.Doges.FirstOrDefaultAsync(d => d.Id == request.DogId, cancellationToken);
+
+            if (dog == null)
+            {
+                return new DeleteDogCommandResult
+                {
+                    DeletingIsSuccessful = false,
+                    Comment = "Dog not found"
+                };
+            }
 
             _ = _dbContext.Doges.Attach(dog);
             _ = _dbContext.Doges.Remove(dog);
